refactor: move tg to services
