+++ app/services/tg/bot/kind/god/startup.py
