+++ app/services/tg/bot/pool.py
@@ -11,15 +11,15 @@ from aiogram.webhook.aiohttp_server import (
 )
 from aiohttp import web
 
+from app.services.tg.bot.kind.ai_relayer.router import general_router
+from app.services.tg.bot.kind.god.router import god_router
+from app.services.tg.bot.kind.god.startup import GOD_BOT_PATH, GOD_BOT_TOKEN, on_startup
+from app.services.tg.bot.types.agent import BotPoolAgentItem
+from app.services.tg.bot.types.bot import BotPoolItem
+from app.services.tg.bot.types.kind import Kind
+from app.services.tg.bot.types.router_obj import RouterObj
+from app.services.tg.utils.cleanup import clean_token_str
 from models.agent import Agent
-from tg.bot.kind.ai_relayer.router import general_router
-from tg.bot.kind.god.router import god_router
-from tg.bot.kind.god.startup import GOD_BOT_PATH, GOD_BOT_TOKEN, on_startup
-from tg.bot.types.agent import BotPoolAgentItem
-from tg.bot.types.bot import BotPoolItem
-from tg.bot.types.kind import Kind
-from tg.bot.types.router_obj import RouterObj
-from tg.utils.cleanup import clean_token_str
 
 logger = logging.getLogger(__name__)
